Fix Headings answers content
Before, if heading was asked, only the questions under it where in answers and often there could be none question under headings. Now in answer, there are also all elemnts before the first question. So that if the notes are structured as question in heading and than some texts or links, this important information isn't excluded in the questions file.
	modified:   quizGenerator/fileManagerClasses.cs