Fix 'Error on Answer press on the end of quiz' Bug
When the user goes throw all questions, there is a blank page and the option to continue. If in this state he presses Answer button, there is no question to open answer for and so an Error occures.
In this commit, the Bug is fixed, so that the Answer button in this state stops working and does nothing.
	modified:   quizGenerator/questionsForm.cs