Add function for looking for prerequisite questions
This function in 'createQuestionsText' function in each new question looks for question tag in '[]' brackets, and stores its insides as the name of the question in a list with its question id. Subsequent question (not yet added) should in in new question look for this tag and on find add the question id to the list of its prerequisites.